all books page books  are display perfectly

+++ src/Pages/BorrowedBooks.jsx
@@ -4,6 +4,7 @@ import { useLoaderData } from "react-router-dom";
 import useAuth from "../hooks/useAuth";
 import { useEffect, useState } from "react";
 import axios, { all } from "axios";
+import Swal from "sweetalert2";
 
 
 const BorrowedBooks = () => {
@@ -21,7 +22,7 @@ const BorrowedBooks = () => {
 
     // get book for display
     // useEffect( () => {
-    //     axios.get(`http://localhost:5008/api/v1/borrow-books/${currentUserEmail}`)
+    //     axios.get(`https://e-readers-server.vercel.app/api/v1/borrow-books/${currentUserEmail}`)
     //     .then(res => setAllBook(res.data))
     //     .catch(err => console.log(err.message))
     // }, [currentUserEmail])
@@ -75,7 +76,7 @@ const BorrowedBooks = () => {
         const newObj = { id, author_name, category_name, image, long_description, name, rating, short_description, decreaseBook, recentArrDate, return_data, userName, userEmail }
 
         // put
-        fetch(`http://localhost:5008/api/v1/category-books/category_name/${_id}`, {
+        fetch(`https://e-readers-server.vercel.app/api/v1/category-books/category_name/${_id}`, {
             method: 'PUT',
             headers: {
                 'content-type': 'application/json'
@@ -89,7 +90,17 @@ const BorrowedBooks = () => {
 
 
         // delete
-        fetch(`http://localhost:5008/api/v1/borrow-books/${_id}`, {
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, return it!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://e-readers-server.vercel.app/api/v1/borrow-books/${_id}`, {
             method: 'DELETE'
         })
             .then(res => res.json())
@@ -98,6 +109,14 @@ const BorrowedBooks = () => {
                 const remainigBook = allBook.filter(book => book._id !== _id)
                 setAllBook(remainigBook)
             })
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your book has been returned.",
+                icon: "success"
+              });
+            }
+          });
+        
     }
     // console.log(newBook);
     console.log(allBook);
